+++ Sources/Sentry/SentryNetworkSwizzling.m
@@ -1,6 +1,7 @@
 #import "SentryNetworkSwizzling.h"
 #import "SentryNetworkTracker.h"
 #import "SentrySwizzle.h"
+#import <objc/runtime.h>
 
 @implementation SentryNetworkSwizzling
 
@@ -34,12 +35,21 @@
 
 + (void)swizzleNSURLSessionConfiguration
 {
-    SEL httpAdditionalHeadersSelector = NSSelectorFromString(@"HTTPAdditionalHeaders");
-    SentrySwizzleInstanceMethod(NSURLSessionConfiguration.class, httpAdditionalHeadersSelector,
-        SentrySWReturnType(NSDictionary *), SentrySWArguments(), SentrySWReplacement({
-            return [SentryNetworkTracker.sharedInstance addTraceHeader:SentrySWCallOriginal()];
-        }),
-        SentrySwizzleModeOncePerClassAndSuperclasses, (void *)httpAdditionalHeadersSelector);
+    // iOS 13 doesn't have a method for HTTPAdditionalHeaders. Instead, it only has a property.
+    // Therefore, we need to make sure that NSURLSessionConfiguration has this method to be able to
+    // swizzle it. Otherwise, we would crash. Cause we can't swizzle properties currently, we only
+    // swizzle when the method is available.
+    SEL selector = NSSelectorFromString(@"HTTPAdditionalHeaders");
+    Class classToSwizzle = NSURLSessionConfiguration.class;
+    Method method = class_getInstanceMethod(classToSwizzle, selector);
+
+    if (method != nil) {
+        SentrySwizzleInstanceMethod(classToSwizzle, selector, SentrySWReturnType(NSDictionary *),
+            SentrySWArguments(), SentrySWReplacement({
+                return [SentryNetworkTracker.sharedInstance addTraceHeader:SentrySWCallOriginal()];
+            }),
+            SentrySwizzleModeOncePerClassAndSuperclasses, (void *)selector);
+    }
 }
 
 #pragma clang diagnostic pop